requirements.txt
flet==0.28.3
sqlalchemy==2.0.44
opencv-python==4.12.0.88
numpy==2.2.6
watchdog==6.0.0
httpx==0.28.1
oauthlib==3.3.1
repath==0.9.0
six==1.17.0
anyio==4.11.0
certifi==2025.10.5
httpcore==1.0.9
idna==3.11
h11==0.16.0
sniffio==1.3.1
pyinstaller==6.16.0
pyserial==3.5

